src/products/repeater/0x03.parser.ts
namespace codec {

    /**
     * Repeater 0x02 frame parser
     */
    export class Repeater0x03Parser implements FrameParser {

        readonly deviceType = 'repeater';
        readonly frameCode = 0x03;

        public parseFrame(payload: Buffer, configuration: Buffer, network: Network) {
            const appContent: Content = {
                type: '0x03 Repeater DL confirmation',
                ...this.getDataFromPayload(payload)
            };

            return appContent;
        }

        public getDataFromPayload(payload: Buffer) {
            const appContent: Content = {};
            RepeaterHelper.getUPStatusFromPayload(payload, appContent);
            appContent['downlinkCode'] = RepeaterHelper.getDownlinkDescriptionForCode(payload.readUInt8(2));
            appContent['downlinkErrorCode'] = RepeaterHelper.getErrorDescriptionForCode(payload.readUInt8(3));
            return appContent;
        }

    }

}

